# specs/ticket-service.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticket Service API",
    "description": "Спецификация сервиса управления билетами и сервиса бронирования.\nВсе параметры операций передаются в URL. В POST, PUT и PATCH данные Ticket\nпередаются в query-параметре `ticket` как JSON.\nПоля id, creationDate и event.id генерируются сервером.\n",
    "version": "1.0.0",
    "contact": {
      "name": "Nostya && Dasha"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Основной сервис управления коллекцией билетов"
    }
  ],
...
  "paths": {
    "/tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получить список билетов",
        "description": "Возвращает массив билетов",
        "operationId": "getTickets",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/size"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Массив билетов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  }
                }
              }
            }
          },
          "422": {
            "description": "page и size должны быть целыми числами в допустимых диапазонах; sort и filter должны содержать поддерживаемые поля, направления и операторы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "size должен быть от 1 до 100; получено 101.",
                  "path": "/api/v1/tickets",
                  "violations": [
                    {
                      "field": "size",
                      "message": "size должен быть от 1 до 100; получено 101."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Добавить новый билет",
        "description": "Создаёт новый объект Ticket и добавляет его в коллекцию",
        "operationId": "createTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/ticket"
          }
        ],
        "responses": {
          "201": {
            "description": "Билет успешно создан",
            "headers": {
              "Location": {
                "description": "URL созданного билета",
                "schema": {
                  "type": "string"
                },
                "example": "/api/v1/tickets/10"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Сервер не смог разобрать запрос: обязательный query-параметр ticket отсутствует, содержит некорректное percent-encoding или не является синтаксически корректным JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "В значении ticket после последнего свойства отсутствует закрывающая фигурная скобка.",
                  "path": "/api/v1/tickets",
                  "violations": [
                    {
                      "field": "ticket",
                      "message": "В значении ticket после последнего свойства отсутствует закрывающая фигурная скобка."
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Обязательные поля Ticket нарушены их типы, диапазоны, значения enum или максимальная длина comment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "price должен быть числом больше 0; получено -10.",
                  "path": "/api/v1/tickets",
                  "violations": [
                    {
                      "field": "ticket.price",
                      "message": "price должен быть числом больше 0; получено -10."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "options": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получить разрешённые методы коллекции",
        "description": "Возвращает Allow со списком методов для /tickets. Это описание возможностей ресурса, не настройка CORS.",
        "operationId": "optionsTickets",
        "responses": {
          "204": {
            "description": "Методы перечислены в Allow; тело отсутствует.",
            "headers": {
              "Allow": {
                "description": "HTTP-методы, поддерживаемые ресурсом /tickets.",
                "schema": {
                  "type": "string"
                },
                "example": "GET, POST, OPTIONS"
              }
            }
          }
        }
      }
    },
    "/tickets/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Уникальный идентификатор билета",
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "example": 10
        }
      ],
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получить билет по идентификатору",
        "description": "Возвращает билет по идентификатору. Если значение If-None-Match\nсовпадает с текущим ETag билета, сервис возвращает 304 без тела.\n",
        "operationId": "getTicketById",
        "responses": {
          "200": {
            "description": "Билет успешно найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "Тег текущего представления билета. Меняется при изменении представления.",
                "schema": {
                  "type": "string"
                },
                "example": "\"ticket-10-v1\""
              }
            }
          },
          "304": {
            "description": "Билет не изменился с момента предыдущего запроса",
            "headers": {
              "ETag": {
                "description": "Версия текущего представления билета",
                "schema": {
                  "type": "string"
                },
                "example": "\"ticket-10-v1\""
              }
            }
          },
          "404": {
            "description": "В коллекции билетов отсутствует билет с указанным идентификатором.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Билет с id 10 не найден.",
                  "path": "/api/v1/tickets/10"
                }
              }
            }
          },
          "422": {
            "description": "Параметр id должен быть целым числом int64 больше 0.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "id должен быть целым числом больше 0; получено -1.",
                  "path": "/api/v1/tickets/-1",
                  "violations": [
                    {
                      "field": "id",
                      "message": "id должен быть целым числом больше 0; получено -1."
                    }
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ifNoneMatch"
          }
        ]
      },
      "put": {
        "tags": [
          "Tickets"
        ],
        "summary": "Обновить билет",
        "description": "Полностью обновляет изменяемые поля билета с указанным идентификатором.\nid и creationDate сохраняются. Если поля event совпадают с текущими,\nevent.id сохраняется; иначе сервер создаёт мероприятие с новым event.id.\nЕсли event отсутствует или равен null, связь удаляется.\n",
        "operationId": "updateTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/ticket"
          }
        ],
        "responses": {
          "200": {
            "description": "Билет успешно обновлён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Сервер не смог разобрать запрос: path-параметр id невозможно преобразовать в int64 либо обязательный query-параметр ticket отсутствует, содержит некорректное percent-encoding или не является синтаксически корректным JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "В значении ticket после последнего свойства отсутствует закрывающая фигурная скобка.",
                  "path": "/api/v1/tickets/10",
                  "violations": [
                    {
                      "field": "ticket",
                      "message": "В значении ticket после последнего свойства отсутствует закрывающая фигурная скобка."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "В коллекции билетов отсутствует билет с указанным идентификатором.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Билет с id 10 не найден.",
                  "path": "/api/v1/tickets/10"
                }
              }
            }
          },
          "422": {
            "description": "id должен быть целым int64 больше 0, а поля Ticket должны соответствовать ограничениям схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "price должен быть числом больше 0; получено -10.",
                  "path": "/api/v1/tickets",
                  "violations": [
                    {
                      "field": "ticket.price",
                      "message": "price должен быть числом больше 0; получено -10."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tickets"
        ],
        "summary": "Частично обновить билет",
        "description": "Обновляет только переданные поля билета. Остальные поля сохраняют прежние значения. Поля id, creationDate и event.id изменять нельзя.",
        "operationId": "patchTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/ticketPatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Билет успешно частично обновлён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Сервер не смог разобрать запрос: path-параметр id невозможно преобразовать в int64 либо обязательный query-параметр ticket отсутствует, содержит некорректное percent-encoding или не является синтаксически корректным JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "Параметр ticket содержит неразбираемый JSON.",
                  "path": "/api/v1/tickets/10",
                  "violations": [
                    {
                      "field": "ticket",
                      "message": "Параметр ticket содержит неразбираемый JSON."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "В коллекции билетов отсутствует билет с указанным идентификатором.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Билет с id 10 не найден.",
                  "path": "/api/v1/tickets/10"
                }
              }
            }
          },
          "422": {
            "description": "id должен быть целым int64 больше 0, а переданные поля должны соответствовать схеме TicketPatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "price должен быть числом больше 0; получено -10.",
                  "path": "/api/v1/tickets/10",
                  "violations": [
                    {
                      "field": "ticket.price",
                      "message": "price должен быть числом больше 0; получено -10."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tickets"
        ],
        "summary": "Удалить билет",
        "description": "Удаляет билет с указанным id. Если билет уже удалён или никогда не существовал, также возвращает 204 без тела. Повторный вызов сохраняет состояние отсутствия билета.",
        "operationId": "deleteTicket",
        "responses": {
          "204": {
            "description": "Билет отсутствует в коллекции: удалён этим запросом или отсутствовал до запроса. Тело ответа отсутствует."
          },
          "400": {
            "description": "Сервер не смог разобрать запрос: path-параметр id невозможно преобразовать в int64, например, передано буквенное или дробное значение либо число за пределами int64.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "Параметр id невозможно преобразовать в int64; получено abc.",
                  "path": "/api/v1/tickets/abc",
                  "violations": [
                    {
                      "field": "id",
                      "message": "Параметр id невозможно преобразовать в int64; получено abc."
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Параметр id должен быть целым числом int64 больше 0.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "21.09.2026 14:30:00",
                  "status": 422,
                  "error": "Unprocessable Content",
                  "message": "id должен быть целым числом больше 0; получено -1.",
                  "path": "/api/v1/tickets/-1",
                  "violations": [
                    {
                      "field": "id",
                      "message": "id должен быть целым числом больше 0; получено -1."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ticket": {
        "type": "object",
        "description": "Билет",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "readOnly": true,
            "description": "Уникальный автоматически генерируемый идентификатор билета",
            "example": 10
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Название билета. Строка не может быть пустой",
            "example": "Билет в партер"
          },
          "coordinates": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "creationDate": {
            "type": "string",
            "pattern": "^(0[1-9]|[12][0-9]|3[01])\\.(0[1-9]|1[0-2])\\.\\d{4} ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.\\d{1,9})?$",
            "readOnly": true,
            "description": "Дата и время создания, генерируются сервером. Формат dd.MM.yyyy HH:mm:ss с необязательными долями секунды (1–9 цифр), без часового пояса. Пример: 21 сентября 2026, 14:30. Календарную корректность проверяет сервер.",
            "example": "21.09.2026 14:30:00"
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Цена билета. Значение должно быть больше 0",
            "example": 2500.5
          },
          "comment": {
            "type": "string",
            "maxLength": 341,
            "description": "Комментарий к билету. Поле не может быть null",
            "example": "Место рядом со сценой"
          },
          "type": {
            "$ref": "#/components/schemas/TicketType"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "required": [
          "id",
          "name",
          "coordinates",
          "creationDate",
          "price",
          "comment",
          "type"
        ],
        "title": "Билет"
      },
...
      "Coordinates": {
        "type": "object",
        "description": "Координаты билета",
        "properties": {
          "x": {
            "type": "number",
            "format": "double",
            "minimum": -239,
            "exclusiveMinimum": true,
            "description": "Координата X (> -239)",
            "example": 10.5
          },
          "y": {
            "type": "integer",
            "format": "int64",
            "minimum": -495,
            "exclusiveMinimum": true,
            "description": "Координата Y (> -495)",
            "example": 20
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "Координаты"
      },
      "Event": {
        "type": "object",
        "nullable": true,
        "description": "Мероприятие, для которого предназначен билет",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "readOnly": true,
            "description": "Уникальный автоматически генерируемый идентификатор мероприятия, должен быть >0",
            "example": 7
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Название мероприятия. Строка не может быть пустой",
            "example": "Вечер оперы"
          },
          "ticketsCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "description": "Количество билетов. Значение должно быть больше 0",
            "example": 120
          },
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "required": [
          "id",
          "name",
          "ticketsCount"
        ],
        "title": "Мероприятие"
      },
      "TicketType": {
        "type": "string",
        "description": "Тип билета. Порядок сравнения: VIP < USUAL < BUDGETARY < CHEAP",
        "enum": [
          "VIP",
          "USUAL",
          "BUDGETARY",
          "CHEAP"
        ],
        "title": "Тип билета"
      },
      "EventType": {
        "type": "string",
        "nullable": true,
        "description": "Тип мероприятия",
        "enum": [
          "CONCERT",
          "BASKETBALL",
          "OPERA",
          null
        ],
        "title": "Тип мероприятия"
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Информация об ошибке выполнения запроса",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Время ошибки в UTC в формате dd.MM.yyyy HH:mm:ss, необязательные доли секунды.",
            "pattern": "^(0[1-9]|[12][0-9]|3[01])\\.(0[1-9]|1[0-2])\\.\\d{4} ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\\.\\d{1,9})?$",
            "example": "21.09.2026 14:30:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP-код ошибки",
            "minimum": 400,
            "maximum": 599,
            "example": 422
          },
          "error": {
            "type": "string",
            "description": "Краткое название ошибки",
            "example": "Unprocessable Content"
          },
          "message": {
            "type": "string",
            "description": "Подробное описание причины ошибки",
            "example": "price должен быть больше 0; получено -10."
          },
          "path": {
            "type": "string",
            "description": "URL запроса, при обработке которого произошла ошибка",
            "example": "/api/v1/tickets"
          },
          "violations": {
            "type": "array",
            "description": "Список нарушений схемы с именами полей. Для ошибок разбора может указываться весь параметр ticket.",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Путь поля или имя параметра.",
                  "example": "ticket.price"
                },
                "message": {
                  "type": "string",
                  "description": "Условие, которое нарушено, и переданное значение.",
                  "example": "price должен быть больше 0; получено -10."
                }
              }
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "message",
          "error",
          "path"
        ],
        "title": "Ошибка API"
      },
...
    },
    "parameters": {
      "ticket": {
        "name": "ticket",
        "in": "query",
        "required": true,
        "description": "JSON-объект с изменяемыми полями Ticket, закодированный в URL.\nid, creationDate и event.id не передаются (генерируются сервером).\nevent может отсутствовать или быть null. Все остальные изменяемые поля обязательны.\nSwagger UI сериализует JSON и кодирует значение параметра автоматически.\n",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TicketInput"
            },
            "example": {
              "name": "Билет в партер",
              "coordinates": {
                "x": 10.5,
                "y": 20
              },
              "price": 2500.5,
              "comment": "Место рядом со сценой",
              "type": "VIP",
              "event": null
            }
          }
        }
      },
      "ticketPatch": {
        "name": "ticket",
        "in": "query",
        "required": true,
        "description": "JSON-объект с полями Ticket, которые требуется изменить, закодированный в URL.\nНеобходимо передать хотя бы одно поле. id, creationDate и event.id изменять нельзя.\n",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TicketPatch"
            },
            "example": {
              "price": 3000,
              "comment": "Новое место"
            }
          }
        }
      },
      "page": {
        "name": "page",
        "in": "query",
        "description": "Номер страницы с 0. Фильтрация и сортировка выполняются до пагинации. За пределами выборки возвращается пустой массив.",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "default": 0,
          "minimum": 0
        }
      },
      "size": {
        "name": "size",
        "in": "query",
        "description": "Максимальное количество элементов на странице.",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        }
      },
      "sort": {
        "name": "sort",
        "in": "query",
        "description": "Сортировка задаётся в формате `поле,направление`, где направление —\n`asc` или `desc`. Для сортировки по нескольким полям параметр\nпередаётся несколько раз: `?sort=price,desc&sort=name,asc`.\nВложенные поля записываются через точку, например `coordinates.x`\nили `event.name`.\n",
        "required": false,
        "style": "form",
        "explode": true,
        "schema": {
          "type": "array",
          "minItems": 1,
          "items": {
            "type": "string"
          }
        },
        "examples": {
          "multiple": {
            "summary": "Сначала цена по убыванию, затем название",
            "value": [
              "price,desc",
              "name,asc"
            ]
          }
        }
      },
      "filter": {
        "name": "filter",
        "in": "query",
        "required": false,
        "description": "Условие задаётся в формате `поле:оператор:значение`.\nНесколько параметров filter объединяются операцией AND.\nПоддерживаются операторы `eq`, `ne`, `gt`, `gte`, `lt`, `lte`,\n`contains` и `isnull`. Вложенные поля записываются через точку.\nПример: `?filter=price:gte:100&filter=type:eq:VIP`.\n",
        "style": "form",
        "explode": true,
        "schema": {
          "type": "array",
          "minItems": 1,
          "items": {
            "type": "string"
          }
        },
        "example": [
          "price:gte:100",
          "type:gt:USUAL"
        ]
      },
      "ifNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "description": "ETag, полученный в предыдущем ответе",
        "schema": {
          "type": "string"
        },
        "example": "\"ticket-10-v1\""
      }
    }
  }
}